% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSSTARCH.R
\name{MSSTARCH}
\alias{MSSTARCH}
\title{This function estimates the parameters of n locations with T observations in 2 regimes. The smoothed probabilities
the regimes are also generated using the estimated parameters.}
\usage{
MSSTARCH(Y_i, Wn)
}
\arguments{
\item{Y_i}{A \eqn{T \times n} matrix consisting \eqn{T} observations of \eqn{n} spatial locations.}

\item{Wn}{A \eqn{n \times n} spatial weight matrix.}
}
\value{
\item{df.est}{A data frame containing the estimated parameters, associated standard errors, Z-values, and p-values.
See details for the descriptions for each parameter.}
\item{smoothed.prob}{A matrix containing the estimated smoothed probabilities for 2 regimes.}
}
\description{
This function estimates the parameters of n locations with T observations in 2 regimes. The smoothed probabilities
the regimes are also generated using the estimated parameters.
}
\details{
df.est contains the following parameters:
\enumerate{
\item lambda1/2: Estimated spatial parameter of the 1st/2nd regime.
\item gamma1/2: Estimated temporal parameter of the 1st/2nd regime.
\item rho1/2: Estimated spatio-temporal parameter of the 1st/2nd regime.
\item p/q: Estimated staying probability of the 1st/2nd regime.
\item a1/a2: Estimated homogeneous fixed effect for regime 1 for \eqn{n} spatial locations.
\item s1: Estimated of the variance in the assumed i.i.d. Gaussian variable in the quasi-MLE.
}
}
\examples{
# Example 1:
# Define variables
n      = 36
T1     = 200
lamb1  = 0.2
gamma1 = 0.2
rho1   = 0.2
alpha1 = 0.1
lamb2  = 0.2
gamma2 = 0.8
rho2   = -0.2
alpha2 = 0.1
sigma  = 1
trans_probs = c(0.97,0.93)

# Simulate a n times n queen-contiguity weight matrix

Wn = spdep::nb2mat(spdep::cell2nb(sqrt(n), sqrt(n), type = "queen"))

# Generate the simulated data.
Y  =  sim_tworegdat(n,T1,Wn,c(lamb1,gamma1,rho1,alpha1), c(lamb2,gamma2,rho2,alpha2), sigma, trans_probs) 
Res = MSSTARCH(t(Y[[1]]), Wn)

# Example 2:
Y              <- stock_exchanges_28  # the dataset in this package
Date_ymd       <- rownames(stock_exchanges_28)

# Calculate the weight matrix using the Piccolo distance
# k-nearest neighbors (k=5) along with AR.PIC
k              <- 5 
dist_mat       <- as.matrix(TSclust::diss(t(log(stock_exchanges_28^2)), "AR.PIC"))
Wmat           <- t(sapply(1:ncol(dist_mat), function(i) ifelse(dist_mat[i,] < sort(dist_mat[i,])[k+2] & dist_mat[i,] > 0, 1/k, 0))) 
# k+2, because of the diagonal zero entry and the strict inequality
diag(Wmat)     <- 0

Res            <- MSSTARCH(stock_exchanges_28, Wmat)

# Plot the smoothed probability of regime 1. 

smoothed       <- Res[[2]]
Date           <- as.Date(Date_ymd)
smoothed.reg1  <- smoothed[1:dim(stock_exchanges_28)[1],2]
df             <- data.frame(Date,smoothed.reg1)
ggplot2::ggplot(df, ggplot2::aes(Date,smoothed.reg1)) +       
    ggplot2::geom_point(col="black") +    
    ggplot2::scale_x_date(date_breaks = "year", date_labels = "\%b\n\%Y") +
    ggplot2::geom_line(col="black") + ggplot2::ylab("Smoothed Probability of Regime 1")+
    ggplot2::theme(panel.grid.major = ggplot2::element_blank(), panel.grid.minor = ggplot2::element_blank(),
                panel.background = ggplot2::element_blank(), axis.line = ggplot2::element_line(colour = "black"))      
}
