% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tworegdat.R
\name{sim_tworegdat}
\alias{sim_tworegdat}
\title{This functions simulate the data of n locations in a period of T1 which have a Markovian structure of 2 regimes.}
\usage{
sim_tworegdat(n, T1, Wn, param1, param2, sigma2, trans_probs)
}
\arguments{
\item{n}{An integer. The number of spatial locations}

\item{T1}{An integer. The length of the period considered.}

\item{Wn}{A \eqn{n \times n} matrix. Spatial weight matrix.}

\item{param1}{A vector of length 4. param1 should contains lambda1, gamma1, rho1, alpha1. Details is described in the function "MSSTARCH".}

\item{param2}{A vector of length 4. param2 should contains lambda2, gamma2, rho2, alpha2. Details is described in the function "MSSTARCH".}

\item{sigma2}{A positive real number. The variance of the mean-zero Gaussian error variable.}

\item{trans_probs}{A vector of length 2. This vector contains the staying probability of the 2 states.}
}
\value{
\item{Y}{A \eqn{n \times T1} matrix. This matrix consists of simulated data of n locations in a period of T1.}
\item{out}{A vector of length T1. This vector consists of realized states at each t in 1,2,...,T1.}
}
\description{
This functions simulate the data of n locations in a period of T1 which have a Markovian structure of 2 regimes.
}
\examples{
# Define variables

n      = 36
T1     = 200
lamb1  = 0.2
gamma1 = 0.2
rho1   = 0.2
alpha1 = 0.1
lamb2  = 0.2
gamma2 = 0.8
rho2   = -0.2
alpha2 = 0.1
sigma  = 1
trans_probs = c(0.97,0.93)

# Simulate a n times n queen-contiguity weight matrix

Wn = spdep::nb2mat(spdep::cell2nb(sqrt(n), sqrt(n), type = "queen"))

# Generate the simulated data.
Y  =  sim_tworegdat(n,T1,Wn,c(lamb1,gamma1,rho1,alpha1), c(lamb2,gamma2,rho2,alpha2), sigma, trans_probs) 
}
