% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARCH.R
\name{STARCH}
\alias{STARCH}
\title{QMLE estimation of spatio-temporal data with 1 regime.}
\usage{
STARCH(Y_i, Wn)
}
\arguments{
\item{Y_i}{A \eqn{T \times n} matrix consisting \eqn{T} observations of \eqn{n} spatial locations.}

\item{Wn}{A \eqn{n \times n} spatial weight matrix.}
}
\value{
\item{df.est}{A data frame containing the estimated parameters, associated standard errors, z-values, and p-values.
See details for the descriptions for each parameter.}
}
\description{
QMLE estimation of spatio-temporal data with 1 regime.
}
\details{
df.est contains the following parameters:
\enumerate{
\item lambda: Estimated spatial parameter.
\item gamma: Estimated temporal parameter.
\item rho: Estimated spatio-temporal parameter.
\item a1: Estimated homogeneous fixed effect for \eqn{n} spatial locations.
\item s1: Estimated of the variance in the assumed i.i.d. Gaussian variable in the quasi-MLE.
}
}
\examples{
Y              <- stock_exchanges_28  # the dataset in this package
Date_ymd       <- rownames(stock_exchanges_28)

# Calculate the weight matrix using the Piccolo distance
# k-nearest neighbors (k=5) along with AR.PIC
k              <- 5 
dist_mat       <- as.matrix(TSclust::diss(t(log(stock_exchanges_28^2)), "AR.PIC"))
Wmat           <- t(sapply(1:ncol(dist_mat), function(i) ifelse(dist_mat[i,] < sort(dist_mat[i,])[k+2] & dist_mat[i,] > 0, 1/k, 0))) 
# k+2, because of the diagonal zero entry and the strict inequality
diag(Wmat)     <- 0

Res            <- STARCH(stock_exchanges_28, Wmat)
}
