% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_exchanges_28.R
\docType{data}
\name{stock_exchanges_28}
\alias{stock_exchanges_28}
\title{This data consists of log returns of 28 stock exchanges (26 Asian stock exchanges + NASDAP and NYSE)
from the end of 2013 to the beginning of 2018}
\format{
\subsection{\code{stock_exchanges_28}}{

A matrix with 1100 rows and 28 columns:
\describe{
\item{row}{Date in ymd}
\item{column}{name of stock exchanges}
}
}
}
\source{
\url{https://www.investing.com}
}
\usage{
stock_exchanges_28
}
\description{
This data consists of log returns of 28 stock exchanges (26 Asian stock exchanges + NASDAP and NYSE)
from the end of 2013 to the beginning of 2018
}
\keyword{datasets}
